project2-grpc-middleware/python-client/requirements.txt
googleapis-common-protos
grpcio
grpcio-tools
protobuf==3.17.3
